Extract metrics in a new component
